Add Retries for Processing Responses, Use Reward Timestamp
Rework the API request functions so their return values are wrapped in
an APIResponse type with Processing, Error, & Success states. Add a
function to unwrap the APIResponse type into the ExceptT transformer.
Modify the request type signatures to operate within mtl typeclasses
instead of a discrete transformer stack.

Modify the StakeReward type & the run function so that we can simply use
the (undocumented) `timestamp` field of a reward object instead of
looking up the timestamp of the reward's slot/block.